requirements.txt
chainlit>=1.0.0
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
tenacity>=8.2.0